refactor(ui): share workflow events viewer logic
Extract the last-execution resolution, loading/empty/error states, and
events tab construction shared by the builder events sidebar and the
embedded workflow artifact into events-shared, removing the duplicated
panel logic between them.
